projects/ecommerce-marketing-funnel/dashboard/requirements.txt
streamlit>=1.33.0
pandas>=2.2.0
plotly>=5.22.0
google-cloud-bigquery>=3.25.0
